+++ src/quacc/calculators/qchem/qchem.py
@@ -24,15 +24,7 @@ if TYPE_CHECKING:
         energy: float  # electronic energy in eV
         forces: NDArray  # forces in eV/A
         hessian: NDArray  # Hessian in eV/A^2/amu
-        enthalpy: float  # total enthalpy in eV
-        entropy: float  # total entropy in eV/K
-        qc_output: dict[
-            str, Any
-        ]  # Output from `pymatgen.io.qchem.outputs.QCOutput.data`
-        qc_input: dict[
-            str, Any
-        ]  # Input from `pymatgen.io.qchem.inputs.QCInput.as_dict()`
-        custodian: dict[str, Any]  # custodian.json file metadata
+        taskdoc: dict[str, Any]  # Output from `emmet.core.qc_tasks.TaskDoc`
 
 
 class QChem(FileIOCalculator):
@@ -42,11 +34,7 @@ class QChem(FileIOCalculator):
         "energy",
         "forces",
         "hessian",
-        "enthalpy",
-        "entropy",
-        "qc_output",
-        "qc_input",
-        "custodian",
+        "taskdoc",
     ]
     results: ClassVar[Results] = {}